Move plot definitions to ./plot.py and add unit tests verifying new import path and old import path still work
